feature: 586 refactor classes to have a clean architecture
